Move scripts into a separate project.
