+++ scripts/.eslintignore
@@ -1,0 +1,6 @@
+# don't ever lint node_modules
+node_modules
+# don't lint nyc coverage output
+coverage
+*.config.js
+.*rc.js
