+++ scripts/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "scripts",
+  "scripts": {},
+  "dependencies": {
+    "dsforest": "1.0.0",
+    "get-audio-duration": "1.0.0",
+    "glob": "7.1.6",
+    "googleapis": "51.0.0",
+    "lodash": "4.17.15"
+  },
+  "devDependencies": {
+    "@types/glob": "7.1.1",
+    "@types/lodash": "4.14.150",
+    "@types/node": "^14.0.1",
+    "eslint-config-prettier": "6.11.0",
+    "eslint": "6.8.0",
+    "prettier": "2.0.5",
+    "typescript": "3.8.3"
+  },
+  "private": true
+}
